re-org of mint test
