+++ .github/workflows/integration-tests.yml
@@ -68,6 +68,7 @@ jobs:
         test_type:
         - "DungeonTransferBlock"
         - "XionSendPlatformFee"
+        - "MintModuleNoInflationNoFees"
 
     steps:
       - name: Set up Go 1.18
